Sync with upstream

+++ applications/product/groovyScripts/product/promo/ProductPromoCondServices.groovy
@@ -401,7 +401,7 @@ def productOrderHist() {
                 logError("Error calling getOrderedSummaryInformation service for the PPIP_ORST_HIST ProductPromo condition input value: " + ServiceUtil.getErrorMessage(result))
                 return serviceResult
             } else {
-                BigDecimal orderSubTotal = (BigDecimal) serviceResult.get("totalSubRemainingAmount")
+                BigDecimal orderSubTotal = serviceResult.get("totalSubRemainingAmount")
                 BigDecimal orderSubTotalAndCartSubTotal = orderSubTotal.add(cart.getSubTotal())
                 if (Debug.verboseOn()) logVerbose("Doing order history sub-total compare: orderSubTotal=" + orderSubTotal + ", for the last " + monthsToInclude + " months.")
                 compareBase = orderSubTotalAndCartSubTotal.compareTo(new BigDecimal(condValue))
@@ -447,7 +447,7 @@ def productOrderYear() {
                 logError("Error calling getOrderedSummaryInformation service for the PPIP_ORST_YEAR ProductPromo condition input value: " + ServiceUtil.getErrorMessage(result))
                 return serviceResult
             } else {
-                BigDecimal orderSubTotal = (BigDecimal) result.get("totalSubRemainingAmount")
+                BigDecimal orderSubTotal = result.get("totalSubRemainingAmount")
                 if (Debug.verboseOn()) logVerbose("Doing order history sub-total compare: orderSubTotal=" + orderSubTotal + ", for the last " + monthsToInclude + " months.")
                 compareBase = orderSubTotal.compareTo(new BigDecimal((condValue)))
